# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bracket Boss API
servers:
  - url: /api
...
paths:
...
  /tournaments/{tournamentId}/players:
    post:
      operationId: joinTournament
      tags: [players]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerWithToken"
  /tournaments/{tournamentId}/players/shuffle:
    post:
      operationId: shufflePlayers
      tags: [players]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /tournaments/{tournamentId}/players/{playerId}:
    patch:
      operationId: updatePlayer
      tags: [players]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: removePlayer
      tags: [players]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, tournamentId, firstName, lastName, seed, joinedAt, eloRating]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        partnerName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        eloRating:
          type: number
        rankTitle:
          type: ["string", "null"]
        rankEmoji:
          type: ["string", "null"]
        seed:
          type: integer
        joinedAt:
          type: string
          format: date-time
    PlayerWithToken:
      type: object
      required: [id, tournamentId, firstName, lastName, seed, joinedAt, eloRating, playerToken]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        partnerName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        eloRating:
          type: number
        rankTitle:
          type: ["string", "null"]
        rankEmoji:
          type: ["string", "null"]
        seed:
          type: integer
        joinedAt:
          type: string
          format: date-time
        playerToken:
          type: string
...
    PlayerInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        partnerName:
          type: string
        teamName:
          type: string
    PlayerUpdate:
      type: object
      required: [playerToken]
      properties:
        playerToken:
          type: string
        teamName:
          type: string
        avatarUrl:
          type: string
...
    HostTokenInput:
      type: object
      required: [hostToken]
      properties:
        hostToken:
          type: string
...